# integration/openapi.json
{"openapi":"3.1.0","info":{"title":"Patient Simulation API","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}},"/scenarios":{"get":{"tags":["scenarios"],"summary":"List Scenarios","operationId":"list_scenarios_scenarios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ScenarioListItem"},"type":"array","title":"Response List Scenarios Scenarios Get"}}}}}}},"/test-categories":{"get":{"tags":["test-categories"],"summary":"List Test Categories","operationId":"list_test_categories_test_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TestCategoryRead"},"type":"array","title":"Response List Test Categories Test Categories Get"}}}}}}},"/test-categories/{category_id}/tests":{"get":{"tags":["test-categories"],"summary":"List Tests By Category","operationId":"list_tests_by_category_test_categories__category_id__tests_get","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","title":"Category Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TestSummary"},"title":"Response List Tests By Category Test Categories  Category Id  Tests Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/scenarios/{scenario_id}/relevant-tests":{"get":{"tags":["test-categories"],"summary":"List Relevant Tests","description":"Tests this scenario has a case-specific finding for (see migration 0004)\n-- a hint for which investigations are actually relevant to the active\ncase, narrower than the full /test-categories/{id}/tests catalog.\n\nPurely additive guidance: POST /sessions/{id}/tests still accepts any\ntest_id from the global catalog regardless of what this returns.","operationId":"list_relevant_tests_scenarios__scenario_id__relevant_tests_get","parameters":[{"name":"scenario_id","in":"path","required":true,"schema":{"type":"integer","title":"Scenario Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TestSummary"},"title":"Response List Relevant Tests Scenarios  Scenario Id  Relevant Tests Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/scenarios/{scenario_id}/questions":{"get":{"tags":["questions"],"summary":"List Questions","operationId":"list_questions_scenarios__scenario_id__questions_get","parameters":[{"name":"scenario_id","in":"path","required":true,"schema":{"type":"integer","title":"Scenario Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionRead"},"title":"Response List Questions Scenarios  Scenario Id  Questions Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions":{"post":{"tags":["sessions"],"summary":"Create Session","operationId":"create_session_sessions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}":{"get":{"tags":["sessions"],"summary":"Get Session Review","operationId":"get_session_review_sessions__session_id__get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionReviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}/tests":{"post":{"tags":["sessions"],"summary":"Order Test","operationId":"order_test_sessions__session_id__tests_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestOrderCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestOrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}/answers":{"post":{"tags":["sessions"],"summary":"Submit Answer","operationId":"submit_answer_sessions__session_id__answers_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}/evaluate":{"post":{"tags":["sessions"],"summary":"Evaluate Session","description":"OSCE-style evaluation of the session so far, generated on demand -- not persisted.\n\n`use_case` is typed as a Union even though app/api/deps.py's\nget_evaluate_session_use_case unconditionally builds the async one in\nproduction, the same reasoning as voice.py's chat_voice/voice_websocket\nroutes: tests override that dependency back to the synchronous\nEvaluateSessionUseCase (wired with the fast/deterministic stub generator)\nfor fixtures that don't want to exercise the async orchestration layer --\nsee tests/conftest.py. Both branches are real, live code.\n\nIn production this always responds status=\"pending\" immediately (score/\nsummary null, empty criteria_breakdown) -- the real LLM-judge result (or\na \"failed\" status with no score) is pushed later over the session's WS\nconnection (WS /ws/voice?session_id=..., see app/infrastructure/ws_hub.py)\nas a `{\"type\": \"evaluation\", ...}` frame with the same field names as\nthis response. A 409 (SessionBusyError) means an evaluation for this\nsession is already being generated -- see EvaluateSessionAsyncUseCase's\nmodule docstring for the concurrency policy.\n\n404 if the session or its scenario doesn't exist; 400 if the scenario has\nno gold_standard set (see MissingGoldStandardError, handled centrally by\napp/api/error_handlers.py's generic DomainError -> 400 mapping) -- both\nraised synchronously, before any \"pending\" response goes out.","operationId":"evaluate_session_sessions__session_id__evaluate_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionEvaluationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}/messages":{"post":{"tags":["chat"],"summary":"Post Message","description":"`message` is a placeholder (status=\"pending\") -- the RAG API backing\nevery reply now (6-48s observed per call) means this always ack's fast\nand generates the real reply in the background; see\napp/application/use_cases/post_message_async.py. The real reply arrives\nlater over the session's WS connection (WS /ws/voice, see\napp/infrastructure/ws_hub.py). A 409 (SessionBusyError) means a reply for\nthis session is already being generated -- see that use case's module\ndocstring for the concurrency policy.","operationId":"post_message_sessions__session_id__messages_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/transcribe":{"post":{"tags":["voice"],"summary":"Transcribe","operationId":"transcribe_transcribe_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_transcribe_transcribe_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TranscribeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{session_id}/chat-voice":{"post":{"tags":["voice"],"summary":"Chat Voice","description":"reply_audio_base64/reply_audio_content_type are null in the response\nin production: STT still runs synchronously (fast), so `transcribed_text`\nis always real, but the reply itself comes back as a \"pending\" placeholder\n(no audio yet) -- the real text and its audio are pushed later over WS\n/ws/voice?session_id=... via app/infrastructure/ws_hub.py, the same push\nchannel described in this module's docstring. (Tests that override\nget_process_voice_chat_use_case to the synchronous ProcessVoiceChatUseCase\nstill get the reply and its audio already final in this response --\nsee this module's docstring.)","operationId":"chat_voice_sessions__session_id__chat_voice_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_chat_voice_sessions__session_id__chat_voice_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatVoiceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnswerCreate":{"properties":{"question_id":{"type":"integer","title":"Question Id"},"choice_id":{"type":"integer","title":"Choice Id"}},"type":"object","required":["question_id","choice_id"],"title":"AnswerCreate"},"AnswerCreateResponse":{"properties":{"id":{"type":"integer","title":"Id"},"question_id":{"type":"integer","title":"Question Id"},"choice_id":{"type":"integer","title":"Choice Id"},"answered_at":{"type":"string","format":"date-time","title":"Answered At"}},"type":"object","required":["id","question_id","choice_id","answered_at"],"title":"AnswerCreateResponse","description":"No `is_correct` here -- submitting an answer must never leak correctness."},"AnswerReviewItem":{"properties":{"id":{"type":"integer","title":"Id"},"question_id":{"type":"integer","title":"Question Id"},"choice_id":{"type":"integer","title":"Choice Id"},"is_correct":{"type":"boolean","title":"Is Correct"},"answered_at":{"type":"string","format":"date-time","title":"Answered At"}},"type":"object","required":["id","question_id","choice_id","is_correct","answered_at"],"title":"AnswerReviewItem","description":"Correctness is only ever exposed here, in the post-hoc session review."},"Body_chat_voice_sessions__session_id__chat_voice_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_chat_voice_sessions__session_id__chat_voice_post"},"Body_transcribe_transcribe_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_transcribe_transcribe_post"},"ChatVoiceResponse":{"properties":{"transcribed_text":{"type":"string","title":"Transcribed Text"},"reply":{"$ref":"#/components/schemas/MessageRead"},"reply_audio_base64":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reply Audio Base64"},"reply_audio_content_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reply Audio Content Type"}},"type":"object","required":["transcribed_text","reply"],"title":"ChatVoiceResponse","description":"reply_audio_base64/reply_audio_content_type are null when `reply` is\nstill pending (production -- see ProcessVoiceChatAsyncUseCase): there is\nno audio to send yet, since there is no finished reply text yet. The real\nreply text and its audio are pushed later over the session's WS\nconnection (see app/infrastructure/ws_hub.py) once generation finishes.\nWith the synchronous \"stub\" test double, `reply` is already \"complete\"\nand both audio fields are always populated, exactly as before this field\nexisted."},"ChoiceRead":{"properties":{"id":{"type":"integer","title":"Id"},"text":{"type":"string","title":"Text"}},"type":"object","required":["id","text"],"title":"ChoiceRead"},"EvaluationCriterionRead":{"properties":{"name":{"type":"string","title":"Name"},"passed":{"type":"boolean","title":"Passed"},"feedback":{"type":"string","title":"Feedback"}},"type":"object","required":["name","passed","feedback"],"title":"EvaluationCriterionRead"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageCreate":{"properties":{"content":{"type":"string","title":"Content"}},"type":"object","required":["content"],"title":"MessageCreate"},"MessageRead":{"properties":{"id":{"type":"integer","title":"Id"},"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"status":{"type":"string","title":"Status","default":"complete"}},"type":"object","required":["id","role","content","created_at"],"title":"MessageRead","description":"`status` is \"complete\" immediately for the synchronous \"stub\"\nPatientReplyGenerator test double -- unchanged behavior from before this\nfield existed. In production, an assistant reply starts \"pending\" and is\nfilled in asynchronously by the RAG-API-backed generator (6-48s observed\nper call) -- see PostMessageAsyncUseCase. A client that doesn't care about\nasync generation can ignore this field entirely."},"OrderedTestReviewItem":{"properties":{"id":{"type":"integer","title":"Id"},"test_id":{"type":"integer","title":"Test Id"},"name":{"type":"string","title":"Name"},"ordered_at":{"type":"string","format":"date-time","title":"Ordered At"}},"type":"object","required":["id","test_id","name","ordered_at"],"title":"OrderedTestReviewItem"},"QuestionRead":{"properties":{"id":{"type":"integer","title":"Id"},"text":{"type":"string","title":"Text"},"choices":{"items":{"$ref":"#/components/schemas/ChoiceRead"},"type":"array","title":"Choices"}},"type":"object","required":["id","text","choices"],"title":"QuestionRead","description":"Deliberately excludes correct_choice_id -- that's the answer key."},"ScenarioListItem":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"ScenarioListItem","description":"Deliberately excludes case_text/gold_standard -- those never leave the server."},"SessionCreate":{"properties":{"scenario_id":{"type":"integer","title":"Scenario Id"}},"type":"object","required":["scenario_id"],"title":"SessionCreate"},"SessionCreateResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"}},"type":"object","required":["session_id"],"title":"SessionCreateResponse"},"SessionEvaluationResponse":{"properties":{"status":{"type":"string","title":"Status","default":"complete"},"score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Score"},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary"},"criteria_breakdown":{"items":{"$ref":"#/components/schemas/EvaluationCriterionRead"},"type":"array","title":"Criteria Breakdown"}},"type":"object","title":"SessionEvaluationResponse","description":"`status` mirrors MessageRead.status's pending/generating/complete/failed\nvocabulary (see app/api/schemas/message.py) -- POST /sessions/{id}/evaluate\nnow acks immediately with status=\"pending\" (score/summary null, an empty\ncriteria_breakdown) while the real LLM-judge call runs in the background\n(app/infrastructure/rag_llm_evaluator.py, tens of seconds per call); the\nreal result (or a \"failed\" status with no score at all -- there is no\nrule-based fallback score) arrives later over the session's WS connection\nas a `{\"type\": \"evaluation\", ...}` frame, same field names as this\nresponse (see EvaluateSessionAsyncUseCase). A client that doesn't care\nabout async generation can still treat \"complete\" the same way it always\nread this response."},"SessionReviewResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"scenario_id":{"type":"integer","title":"Scenario Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"messages":{"items":{"$ref":"#/components/schemas/MessageRead"},"type":"array","title":"Messages"},"ordered_tests":{"items":{"$ref":"#/components/schemas/OrderedTestReviewItem"},"type":"array","title":"Ordered Tests"},"answers":{"items":{"$ref":"#/components/schemas/AnswerReviewItem"},"type":"array","title":"Answers"}},"type":"object","required":["session_id","scenario_id","created_at","messages","ordered_tests","answers"],"title":"SessionReviewResponse"},"TestCategoryRead":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"TestCategoryRead"},"TestOrderCreate":{"properties":{"test_id":{"type":"integer","title":"Test Id"}},"type":"object","required":["test_id"],"title":"TestOrderCreate"},"TestOrderResponse":{"properties":{"ordered_test_id":{"type":"integer","title":"Ordered Test Id"},"test_id":{"type":"integer","title":"Test Id"},"name":{"type":"string","title":"Name"},"result":{"type":"string","title":"Result"},"ordered_at":{"type":"string","format":"date-time","title":"Ordered At"}},"type":"object","required":["ordered_test_id","test_id","name","result","ordered_at"],"title":"TestOrderResponse"},"TestSummary":{"properties":{"id":{"type":"integer","title":"Id"},"category_id":{"type":"integer","title":"Category Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","category_id","name"],"title":"TestSummary","description":"The investigation menu -- no `result` here; that only appears once ordered."},"TranscribeResponse":{"properties":{"text":{"type":"string","title":"Text"}},"type":"object","required":["text"],"title":"TranscribeResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}